+++ demo/src/pages/index.mdx
@@ -1,3 +1,6 @@
-<ColorSwatch />
+---
+title: Baton Rouge
+---
+
 <CurrentWeather />
 <TodaysWeather />
