+++ readme.md
@@ -19,6 +19,7 @@ A weather dashboard [Gatsby theme](https://www.gatsbyjs.org/docs/themes/), compo
 - Geolocation-provided weather
 - Weather context state hook available for custom components - `useWeather`
 - Refresh interval for data fetching
+- Carbon Design System Iconography
 
 ## Installation
