+++ theme/src/hooks/useInterval.js
@@ -1,0 +1,19 @@
+import React from 'react'
+
+const useInterval = (callback, delay) => {
+  const savedCallback = React.useRef()
+  React.useEffect(() => {
+    savedCallback.current = callback
+  }, [callback])
+  React.useEffect(() => {
+    const tick = () => {
+      savedCallback.current()
+    }
+    if (delay !== null) {
+      const id = setInterval(tick, delay)
+      return () => clearInterval(id)
+    }
+  }, [delay])
+}
+
+export default useInterval
